configurations: sbp1: Add MAX5970 SSB
SBP1 uses the MAX5970 eFuse to protect various FRUs, like rulerdrive
SSDs, fans and M.2 slots. While the MAX5970 is an eFuse and not a regulator
it's threated as such by setting a fixed output voltage that is always
equal to the input voltage. In addition to protecting the outputs it
provides voltage and current readings.

Support has been added to dbus-sensors in [1] and the Linux kernel [2].

Tested: All 27 MAX5970 channels provide sane readings.

1: https://gerrit.openbmc.org/c/openbmc/dbus-sensors/+/67549
2: https://github.com/torvalds/linux/commit/f5afdd13ed6c643c7243e685fe3cf5484b3fdfae

Change-Id: Ifb0216335a25215ae464e709e7db7f757e96baf1
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>