configurations: sbp1: Add rulerdrive efuses
SBP1 has 32 slots for rulerdrive SSDs. Every slot has it's own
MAX5970 eFuse that provides voltage and current sensor readings.

While the MAX5970 is on the baseboard it only makes sense to show
the sensor values when a rulerdrive is plugged into the corresponding
slot.
Thus add the MAX5970 as separate chassis object with the probe condition
to evaluate if the slot is occupied.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>