Add backup mechanism for eeproms
This is the follow-up patch for patch #74672 [1], which introduces
manual updating of FRU properties via DBus.

This change adds a simple mechanism to backup eeprom files before
altering the contents of it via UpdateFruProperty dbus method, to add
some safety to the updating.

The mechanism only stores a backup of the original eeprom,
before any manual updating via UpdateFruProperty occured.
A backup is automatically created, whenever the user does a first update
of the FRU data via dbus.
The backups are stored in /var/lib/entity-manager/backups.
The hash calculated from the fru data is used as unique key to
identify the backup file.
A symlink with the hash value of the updated fru data used as key,
points to the original backup file. Subsequent updates with
UpdateFruProperty also update the symlink with the hash of the updated
FRU to maintain the mapping between the FRUDevice and its backup.

In order to restore the FRU, the RestoreFruFromBackup method is exposed
on dbus. This is to be seen as internal debug interface.

Depends on:
https://gerrit.openbmc.org/c/openbmc/openbmc/+/74674

Tested: The change has been tested in qemu manually by calling
UpdateFruProperty and checking if the backup was created and if the
call to RestoreFruFromBackup actually restores to the original contents.

[1] https://gerrit.openbmc.org/c/openbmc/entity-manager/+/74672

Change-Id: If2e33ae2e5d639fc9bad4d5f3eafedbf4cea0679
Signed-off-by: Oliver Brewka <mox669.dev@gmail.com>